added lots of type casting for 64bit compatibility

+++ EQEmuServer/EQEmuLoginServer/SecurityLibrary.cpp
@@ -121,7 +121,7 @@ const char *SecurityLibrary::GetError()
 	unsigned long err = GetLastError();
 	if(err == 0)
 		return(NULL);
-	sprintf(ErrBuf, "Error #%lu", err);
+	sprintf(ErrBuf, "Error #%lu", (unsigned long)err);
 	return(ErrBuf);
 #else
 	return GetLastError();
